pyproject.toml
[tool.poetry]
name = "astro-elvis"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
apache-airflow = {extras = ["amazon"], version = "^2.2.5"}
boto3 = "^1.21.46"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
